+++ second/load.h
@@ -5,4 +5,4 @@
  */
                                                                                 
 extern char* load_kernel(char* path, int bootstrap_size, unsigned long *base, unsigned long *entry, unsigned long *size);
-extern char* load_ramdisk(char* path, int *ramdisk_size);
+extern char* load_ramdisk(char* path, unsigned long *ramdisk_size);
